backend/requirements.txt
fastapi
uvicorn
pydantic
deepface
transformers
pillow
openai-whisper
ffmpeg-python
